requirements.txt
audioread==2.1.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
ibm-cloud-sdk-core==3.15.1
ibm-watson==5.3.1
idna==3.3
joblib==1.1.0
nltk==3.7
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
regex==2022.4.24
requests==2.27.1
six==1.16.0
tqdm==4.64.0
urllib3==1.26.9
websocket-client==1.1.0
